Added three files
There are three files, including a markdown file, an R script and a stan file. The R script includes all the functions that I rewrote and the markdown file includes the code calling these functions. The stan file has the stan code that needs to be compiled when fitting the model.